requirements.txt
bcrypt
databases
fastapi
humanfriendly
python-dotenv
uvicorn[standard]
